+++ datumaro/plugins/icdar_format/extractor.py
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: MIT
 
 from glob import iglob
+import logging as log
 import os.path as osp
 
 import numpy as np
@@ -59,16 +60,13 @@ class _IcdarExtractor(SourceExtractor):
                 objects = line.split(', ')
                 if len(objects) == 2:
                     image = objects[0]
-                    objects = objects[1].split('\"')
-                    if 1 < len(objects):
-                        if len(objects) % 2:
-                            captions = [objects[2 * i + 1]
-                                for i in range(int(len(objects) / 2))]
+                    captions = []
+                    for caption in objects[1:]:
+                        if caption[0] != '\"' or caption[-1] != '\"':
+                            log.warning("Line %s: unexpected number "
+                                "of quotes" % line)
                         else:
-                            raise Exception("Line %s: unexpected number "
-                                "of quotes in filename" % line)
-                    else:
-                        captions = objects[0].split()
+                            captions.append(caption.replace('\\', '')[1:-1])
                 else:
                     image = objects[0][:-1]
                     captions = []